Moved Proposal / Rfp struct into shared lib
