+++ Cargo.toml
@@ -13,12 +13,15 @@ near-contract-standards = "5.1.0"
 serde_json = { version = "1.0", features = ["preserve_order"] }
 devhub_common = { path = "./devhub_common" }
 html-escape = "0.2.13"
+devhub-shared = { path = "shared" }
 
 [dev-dependencies]
 near-sdk = { version = "5.1.0", features = ["unit-testing"] }
 insta = { version = "1.31.0", features = ["json", "redactions"] }
 regex = "1"
-near-workspaces = { version = "0.10.0", features = ["unstable"], default-features = false  }
+near-workspaces = { version = "0.10.0", features = [
+  "unstable",
+], default-features = false }
 tokio = { version = "1.10.0", features = ["full"] }
 anyhow = "1.0"
