+++ shared/.cargo/config.toml
@@ -1,0 +1,1 @@
+[exclude]
