+++ shared/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod proposal;
+pub mod rfp;
+pub mod str_serializers;
