+++ shared/src/str_serializers.rs
